+++ src/components/OutputField.jsx
@@ -1,21 +1,21 @@
 import React, { Component } from 'react';
-import styles               from '../App.module.css';
+import styles               from './UserProfile.module.css';
 
 
 class OutputField extends Component {
 
     render(){
-        let { value } = this.props;
+        let { value }                   = this.props;
         const { className, title = "" } = this.props;
-        const DATE_TITLE = "Дата рождения: ";
+        const DATE_TITLE                = "Дата рождения: ";
 
         if ( title === DATE_TITLE && value ) {
             value = value.slice(8) + "." + value.slice(5,7) + "." + value.slice(0,4)
         }
         return(
-            <div className = { className || styles.stringProfile } >
+            <div className = { styles.containerOutputField } >
                 <div className = { styles.outputFieldTitle }>{ title }</div>
-                <div className = { styles.outputField }>{ value }</div>
+                <div className = { className || styles.outputField }>{ value }</div>
             </div>
         )
     }
