+++ src/components/TextArea.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import styles               from '../App.module.css';
+import styles               from './Form.module.css';
 
 
 class TextArea extends Component {
