+++ src/components/UserProfile.module.css
@@ -1,0 +1,24 @@
+.containerOutputField{
+    margin: 8px 5px;
+    align-self: flex-start;
+    width: 100%;
+}
+.outputFieldTitle{
+    font-size: 14px;
+    color:rgba(22, 3, 30, 0.75);
+    margin:  0;
+    text-decoration: underline;
+}
+.outputField{
+    padding: 3px 0;
+    font-size: 16px;
+    resize: none;
+    width: 100%;
+    border: 0;
+    background-color: #fdfcff;
+}
+.titleForm{
+    font-size: x-large;
+    text-align: center;
+}
+
